+++ R/sig_fit_bootstrap_batch.R
@@ -95,13 +95,17 @@ sig_fit_bootstrap_batch <- function(catalogue_matrix,
       sample = names(expo_list$errors),
       errors = as.numeric(expo_list$errors)
     )
+    expo_list$cosine <- data.table::data.table(
+      sample = names(expo_list$cosine),
+      cosine = as.numeric(expo_list$cosine)
+    )
     optimal_list[[m]] <- expo_list
   }
   optimal_list <- purrr::transpose(optimal_list)
   optimal_list <- purrr::map(optimal_list, data.table::rbindlist, fill = TRUE, idcol = "method")
 
   ## Get bootstrap exposures with different methods
-  send_info("Getting bootstrap exposures (&errors) for different methods.")
+  send_info("Getting bootstrap exposures (&errors/similarity) for different methods.")
   send_info("This step is time consuming, please be patient.")
   call_bt <- function(x, sample, y, methods, n = 1000, ...) {
     if (!is.null(job_id)) {
@@ -170,6 +174,10 @@ sig_fit_bootstrap_batch <- function(catalogue_matrix,
         type = names(out$errors),
         errors = as.numeric(out$errors)
       )
+      out$cosine <- data.table::data.table(
+        type = names(out$cosine),
+        cosine = as.numeric(out$cosine)
+      )
       out
     })
     x <- purrr::transpose(x)
@@ -201,13 +209,16 @@ sig_fit_bootstrap_batch <- function(catalogue_matrix,
   send_success("Outputing.")
   optimal_list$expo$type <- "optimal"
   optimal_list$errors$type <- "optimal"
+  optimal_list$cosine$type <- "optimal"
 
   expos <- rbind(optimal_list$expo, bt_list$expo, fill = TRUE)
   errors <- rbind(optimal_list$errors, bt_list$errors, fill = TRUE)
+  cosine <- rbind(optimal_list$cosine, bt_list$cosine, fill = TRUE)
 
   result <- list(
     expo = expos,
     error = errors,
+    cosine = cosine,
     p_val = p_val
   )
   return(result)